feat: Fix typo, add Header component, add CoreConcepts component and add styling for Header component
