+++ src/components/CoreConcepts.jsx
@@ -1,0 +1,9 @@
+export default function CoreConcepts({ image, title, description }) {
+  return (
+    <li>
+      <img src={image} alt={title} />
+      <h3>{title}</h3>
+      <p>{description}</p>
+    </li>
+  );
+}
