+++ src/components/Header/Header.jsx
@@ -1,0 +1,19 @@
+import "./Header.css";
+import reactImage from "../../assets/react-core-concepts.png";
+const reactDescriptions = ["Fundamental", "Crucial", "Core"];
+function genRandomInt(max) {
+  return Math.floor(Math.random() * (max + 1));
+}
+export default function Header() {
+  const Description = reactDescriptions[genRandomInt(2)];
+  return (
+    <header>
+      <img src={reactImage} alt="Stylized atom" />
+      <h1>React Essentials</h1>
+      <p>
+        {Description} React concepts you will need for almost any app you are
+        going to build!
+      </p>
+    </header>
+  );
+}
